pyproject.toml
[project]
name = "micro-ml"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "harish gehlot", email = "gehloth03@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "hydra-core>=1.3.2",
    "omegaconf>=2.3.0",
    "polars>=1.40.1",
    "prek>=0.3.10",
    "ruff>=0.15.12",
    "rust-just>=1.50.0",
    "scikit-learn>=1.8.0",
    "skops>=0.14.0",
]

[build-system]
requires = ["uv_build>=0.9.2,<0.10.0"]
build-backend = "uv_build"
